Add files via upload
We study the merger history of primordial black holes (PBHs) in a scenario where they represent the dominant dark matter component of a typical dwarf galaxies' core. We investigate the possibility of a sequence of collisions resulting in a hierarchical merger of black holes and look at the final mass spectrum in such clusters, which initially present a monochromatic (single-mass) PBH population.